[INFRA-231]
jenkins-admin isn't authenticating itself with nickserv, and it turns out the reason for that is because the Java program isn't getting the nick password through command line arguments as it should be.

After a bit of digging, I found out that this is because ENTRYPOINT in the shell form doesn't correctly get the additional arguments given to "docker run" (it runs it like sh -c "java -jar ..." PASSWORD as opposed to sh -c "java -jar ... PASSWORD")

Changing it to the long form to avoid this.

See: https://serverfault.com/questions/647779/