Dont add committers to everyone group by default
When forking a repo of the form

   https://github.com/pouet/therepo-plugin

with the classic:

    fork https://github.com/pouet/therepo-plugin as therepo-plugin

pouet user will automatically be added to the 'Everyone' group.

This is counter-intuitive, IMO, and non consistent with the same usage
when pouet is an org.

You'll end up doing N calls to:

    make someone committer of therepo-plugin

to give her/him access.

Hence this modification makes the behaviour more consistent across those
two cases, and more explicit.